if_rge: flesh out rge_newbuf() and the RX ring setup path
This isn't yet called, but it does compile and it did require the
busdma stuff to be swizzled and the lack of if_rxr (rx ring management)
to be coded around.

The next thing is rge_iff() - the RX filter and multicast stuff, which
rge_init() requires.

We're almost ready to bring up rge_init() and test init/stop.