rge: fix RX ring handling to not re-check descriptors we've cleaned
I .. hilariously wasn't checking that we weren't running outside of the
prod/cons ring so after I stopped allocating a full ring, I noticed
I was crashing on a null pointer dereference.

It turns out that I was re-checking a frame that I had already checked,
which has a NULL mbuf pointer at this point AND the OWN flag in the
descriptor status field wasn't set - so the driver thought the hardware
had finished with it.

Whoops.  Fix that!